requirements.txt
shiny
sentence-transformers
PyMuPDF
faiss-cpu
numpy
scikit-learn
SpeechRecognition
pyttsx3
pyaudio